mem test

+++ A2/src/CalendarParser.c
@@ -2505,7 +2505,7 @@ char *calendarToJSON(const Calendar *cal) {
     }
     tempJSON = calloc(1, sizeof(char) * 200);
     tempJSON = realloc(tempJSON, sizeof(char) * (strlen(cal->prodID)) + 100);
-    sprintf(tempJSON, "{\"version\":%f,\"prodID\":\"%s\",\"numProps\":%d,\"numEvents\":%d}", cal->version,cal->prodID,2+getLength(cal->properties), getLength(cal->events));
+    sprintf(tempJSON, "{\"version\":%d,\"prodID\":\"%s\",\"numProps\":%d,\"numEvents\":%d}", (int)cal->version,cal->prodID,2+getLength(cal->properties), getLength(cal->events));
     printf("JSON CALENDAR -> %s\n", tempJSON);
     eventListToJSON(cal->events);
     deallocator(tempJSON);
